Corrigindo GameLoan

+++ src/main/java/br/ufms/cpcx/api/gamersclub/services/PartnerService.java
@@ -2,6 +2,7 @@ package br.ufms.cpcx.api.gamersclub.services;
 
 import java.util.Optional;
 
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Example;
 import org.springframework.data.domain.ExampleMatcher;
 import org.springframework.data.domain.Page;
@@ -9,7 +10,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import br.ufms.cpcx.api.gamersclub.dtos.PartnerDto;
+import br.ufms.cpcx.api.gamersclub.dtos.PartnerGameDto;
 import br.ufms.cpcx.api.gamersclub.models.PartnerModel;
 import br.ufms.cpcx.api.gamersclub.repositories.PartnerRepository;
 
@@ -17,9 +18,11 @@ import br.ufms.cpcx.api.gamersclub.repositories.PartnerRepository;
 @Service
 public class PartnerService {
     final PartnerRepository partnerRepository;
+    final GameLoanService gameLoanService;
 
-    public PartnerService(PartnerRepository partnerRepository) {
+    public PartnerService(PartnerRepository partnerRepository, GameLoanService gameLoanService) {
         this.partnerRepository = partnerRepository;
+        this.gameLoanService = gameLoanService;
     }
 
     @Transactional
@@ -29,6 +32,8 @@ public class PartnerService {
 
     @Transactional
     public void delete(PartnerModel partnerModel) {
+        if(gameLoanService.checkHasLoan(partnerModel.getId())) throw new DataIntegrityViolationException("There are partner relationships");
+
         partnerRepository.delete(partnerModel);
     }
 
@@ -44,8 +49,8 @@ public class PartnerService {
         return partnerRepository.findByNameAndPhoneNumber(partnerModel.getName(), partnerModel.getPhoneNumber());
     }
 
-    public Page<PartnerModel> findByPartnerFilter(PartnerDto partnerDto, Pageable pageable) {
-        var partnerModel = partnerDto.getPartnerModel();
+    public Page<PartnerModel> findByPartnerFilter(PartnerGameDto partnerGameDto, Pageable pageable) {
+        var partnerModel = partnerGameDto.getPartnerModel();
 
         ExampleMatcher matcher = ExampleMatcher.matching()
                 .withMatcher("name", match -> match.startsWith())
@@ -56,5 +61,4 @@ public class PartnerService {
 
         return partnerRepository.findAll(example, pageable);
     }
-
 }
